Split declare into separate files
