+++ protostar/commands/declare/__init__.py
@@ -1,0 +1,1 @@
+from .declare_command import DeclareCommand
